Вынос emtpySymbol, который используется только для построения SymbolTree, за пределы интерфейса
