+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/context/symbol/SourceDefinedSymbol.java
@@ -22,13 +22,8 @@
 package com.github._1c_syntax.bsl.languageserver.context.symbol;
 
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
-import com.github._1c_syntax.bsl.languageserver.utils.Ranges;
-import lombok.Getter;
-import lombok.Setter;
 import org.eclipse.lsp4j.Range;
-import org.eclipse.lsp4j.SymbolKind;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
 
@@ -47,30 +42,5 @@ public interface SourceDefinedSymbol extends Symbol {
 
   default Optional<SourceDefinedSymbol> getRootParent() {
     return getParent().flatMap(SourceDefinedSymbol::getRootParent).or(() -> Optional.of(this));
-  }
-
-  static SourceDefinedSymbol emptySymbol() {
-    return new SourceDefinedSymbol() {
-      @Getter
-      private final DocumentContext owner = null;
-      @Getter
-      private final String name = "empty";
-      @Getter
-      private final SymbolKind symbolKind = SymbolKind.Null;
-      @Getter
-      private final Range range = Ranges.create(-1, 0, -1, 0);
-      @Getter
-      private final Range selectionRange = Ranges.create(-1, 0, -1, 0);
-      @Getter
-      @Setter
-      private Optional<SourceDefinedSymbol> parent = Optional.empty();
-      @Getter
-      private final List<SourceDefinedSymbol> children = Collections.emptyList();
-
-      @Override
-      public void accept(SymbolTreeVisitor visitor) {
-      }
-
-    };
   }
 }
